% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sp_function.R
\name{compare_sp}
\alias{compare_sp}
\title{Compare species annotations across CAMDA, NCBI, and antibiogram datasets}
\usage{
compare_sp(metadata_db = NULL, sra_metadata_db = NULL, antibiograms_db = NULL)
}
\arguments{
\item{metadata_db}{A data frame containing CAMDA species annotations.
Must include the columns: \code{accession} and \code{scientific_name_CAMDA}.
Example datasets included in the package: \code{training_metadata_db}, \code{testing_metadata_cleaned}.}

\item{sra_metadata_db}{A data frame containing metadata from NCBI.
Must include the columns: \code{accession} and \code{scientific_name_NCBI}.
A sample dataset is preloaded in the package as \code{sra_metadata_db}.}

\item{antibiograms_db}{A data frame containing species annotations from antibiograms.
Must include the columns: \code{accession} and \code{scientific_name_Antibiogram}.
A sample dataset is preloaded in the package as \code{antibiograms_db}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{accession}: Unique sample identifier.
\item \code{scientific_name_CAMDA}: Species annotation from the CAMDA dataset.
\item \code{scientific_name_NCBI}: Species annotation from NCBI metadata.
\item \code{scientific_name_Antibiogram}: Species annotation from antibiogram metadata.
\item \code{status}: Consistency between CAMDA and external sources, with values:
\describe{
\item{Matched_NCBI_Antibiogram}{CAMDA matches both NCBI and antibiogram.}
\item{Matched_NCBI}{CAMDA matches only NCBI.}
\item{Matched_Antibiogram}{CAMDA matches only the antibiogram.}
\item{Mismatch_with_CAMDA}{At least one source has data, but CAMDA does not match either.}
\item{Missing_All}{No data available from NCBI or antibiogram.}
}
\item \code{status_reference}: Consistency between NCBI and antibiogram sources, with values:
\describe{
\item{Good_sources}{Both sources are present and agree.}
\item{Verify_sources}{At least one source is present and sources disagree or are partially missing.}
\item{No_sources}{Both NCBI and antibiogram are missing.}
}
}
}
\description{
This function compares species annotations from the CAMDA dataset against metadata from
NCBI and antibiograms. It returns a dataset that includes consistency status between CAMDA
and the external sources, as well as a comparison between the external sources themselves.
}
\examples{
result <- compare_sp(metadata_db = training_metadata_db, sra_metadata_db, antibiograms_db)

}
